feat: #8 set room for user
