+++ .eslintrc.json
@@ -18,6 +18,7 @@
   "rules": {
     "no-console": [
       "off"
-    ]
+    ],
+    "no-param-reassign": ["error", { "props": false }]
   }
 }
