+++ main.js
@@ -3,8 +3,6 @@ const express = require('express');
 const cors = require('cors');
 const bodyParser = require('body-parser');
 const http = require('http');
-const socketio = require('socket.io');
-const { instrument } = require('@socket.io/admin-ui');
 const mongoose = require('mongoose');
 const userRouter = require('./src/routes/user.route');
 const lobbyRouter = require('./src/routes/lobby.route');
@@ -38,29 +36,11 @@ app.use('/lobbys', lobbyRouter);
 
 const server = http.createServer(app);
 
-const io = socketio(server, { cors: { origin: '*' } });
+const websocket = require('./src/socketHandler/websocket');
 
-const { test } = require('./src/socketHandler/testHandler')(io);
+websocket.createIO(server);
 
-const onConnection = (socket) => {
-  socket.on('test:test', test);
-};
-
-io.on('connection', onConnection);
-
-// SocketIO admin UI
-if (process.env.SIO_ADMINUI_USERNAME && process.env.SIO_ADMINUI_PASSWORD) {
-  app.use('/sioadmin', express.static('./node_modules/@socket.io/admin-ui/ui/dist'));
-
-  instrument(io, {
-    auth: {
-      type: 'basic',
-      username: process.env.SIO_ADMINUI_USERNAME,
-      password: process.env.SIO_ADMINUI_PASSWORD,
-    },
-    mode: 'development',
-  });
-}
+app.use('/sioadmin', express.static('./node_modules/@socket.io/admin-ui/ui/dist'));
 
 server.listen(process.env.PORT, () => {
   console.log(`Server running on port ${process.env.PORT}  http://localhost:${process.env.PORT}`);
