+++ src/services/socket.service.js
@@ -1,0 +1,5 @@
+const { io } = require('../socketHandler/websocket');
+
+exports.joinRoom = async (roomName, socketID) => {
+  (await io().in(socketID).fetchSockets())[0].join(roomName);
+};
